+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to ChamelShot.
 
-## Unreleased
+## v5.1.0 — 2026-08-07
 
 ### Added